fix: remove granular status feature flag
