+++ app/client/src/ce/entities/FeatureFlag.ts
@@ -29,7 +29,6 @@ export const FEATURE_FLAG = {
     "release_git_continuous_delivery_enabled",
   release_git_autocommit_feature_enabled:
     "release_git_autocommit_feature_enabled",
-  release_git_status_granular_enabled: "release_git_status_granular_enabled",
   license_widget_rtl_support_enabled: "license_widget_rtl_support_enabled",
   release_show_partial_import_export_enabled:
     "release_show_partial_import_export_enabled",
@@ -75,7 +74,6 @@ export const DEFAULT_FEATURE_FLAG_VALUE: FeatureFlags = {
   release_drag_drop_building_blocks_enabled: false,
   license_git_branch_protection_enabled: false,
   release_git_autocommit_feature_enabled: false,
-  release_git_status_granular_enabled: false,
   license_git_continuous_delivery_enabled: false,
   release_git_continuous_delivery_enabled: false,
   license_widget_rtl_support_enabled: false,
