+++ app/client/src/pages/Editor/gitSync/components/GitChangesList/index.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { render } from "@testing-library/react";
 import { Provider } from "react-redux";
 import configureStore from "redux-mock-store";
-import GitChangesList from "./NewChangesList";
+import GitChangesList from ".";
 import { merge } from "lodash";
 import type { GitStatusData } from "reducers/uiReducers/gitSyncReducer";
 import "@testing-library/jest-dom";
